requirements.txt
celery==5.3.6
Django==5.0.1
django-admin-interface==0.28.3
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-nested-admin==4.0.2
djangorestframework==3.14.0
django-storages==1.14.2
django-widget-tweaks==1.5.0
psycopg2==2.9.9 --no-binary psycopg2
pur==7.3.1
pylint==3.0.3
pylint-django==2.5.5
pylint-fail-under==0.3.0
pylint-plugin-utils==0.8.2
requests==2.31.0
sentry-sdk==1.39.2
URLObject==2.4.3
wheel==0.42.0
